Update FirstData.php
I added track1 and track2 as options in the post data.  This allowed us to accept input from USB magnetic stripe credit card readers.

We use a USB swiper that has keyboard emulation.  We just capture the text that is read in.  Split the track string by the semi-colon character to split it into track1, track2, and track 3 (First Data only appears to allow for track1 and track2 data though).